+++ backend/src/main/java/reviewme/keyword/repository/KeywordRepository.java
@@ -3,7 +3,7 @@ package reviewme.keyword.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import reviewme.keyword.domain.Keyword;
-import reviewme.keyword.exception.KeywordNotFoundException;
+import reviewme.keyword.domain.exception.KeywordNotFoundException;
 
 @Repository
 public interface KeywordRepository extends JpaRepository<Keyword, Long> {
