+++ backend/src/main/java/reviewme/review/repository/QuestionRepository.java
@@ -3,7 +3,7 @@ package reviewme.review.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import reviewme.question.domain.Question;
-import reviewme.review.exception.QuestionNotFoundException;
+import reviewme.question.domain.exception.QuestionNotFoundException;
 
 @Repository
 public interface QuestionRepository extends JpaRepository<Question, Long> {
