+++ backend/src/main/java/reviewme/review/repository/ReviewRepository.java
@@ -3,7 +3,7 @@ package reviewme.review.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import reviewme.review.domain.Review;
-import reviewme.review.exception.ReviewNotFoundException;
+import reviewme.review.domain.exception.ReviewNotFoundException;
 
 @Repository
 public interface ReviewRepository extends JpaRepository<Review, Long> {
